refactor: move Server to package cc
